requirements.txt
Django==1.10.1
Pillow==3.3.1
PyJWT==1.4.2
argparse==1.2.1
dj-database-url==0.4.1
django-allauth==0.27.0
django-social-auth==0.8.1
gunicorn==19.6.0
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==2.0.0
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==3.2.1
wsgiref==0.1.2
psycopg2==2.5.4
